[212]: UserInfo component refactoring
closes #212
