--- apps/ligretto-frontend/src/components/blocks/home/user-info/UserInfo.module.scss
@@ -1,3 +1,0 @@
-.userInfo {
-  margin-bottom: 1rem;
-}
