--- apps/ligretto-frontend/src/components/blocks/home/user-info/index.ts
@@ -1,1 +1,0 @@
-export { UserInfo } from './UserInfo'
